requirements.txt
accelerate==0.3.0
albumentations==1.0.3
click==8.0.1
flake8==3.9.2
inplace-abn==1.1.0
isort==5.9.3
matplotlib==3.4.2
numpy==1.21.1
ordered-set==4.0.2
pandas==1.3.1
pydantic==1.8.2
pytest==6.2.4
python-dotenv==0.19.0
seaborn==0.11.1
tifffile==2021.7.2
timm==0.4.12
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.61.2
yapf==0.31.0
